nixos/installer: Add comment for imports
Beginners shouldn't be confusing the module system's special attribute `imports` with the primitive
import function, as they may be similar but different purposes, and they
should be provided a resource to get a better understanding of how NixOS
modules work.

This is an attempt to lessen the confusion between the module system and the language itself.